requirements.txt
numpy==2.2.6
pandas
matplotlib==3.10.5
seaborn
pyarrow
fastparquet

# notebook env
ipykernel
ipython
openpyxl

# machine learning
scikit-learn

# Proceesing multi types of data
docling

# RAG requirements
# === LangChain Core & Engine ===
langchain              # Main framework to build LLM-based applications using modular chains and agents
langchain-core         # Lightweight core module: contains schemas, runtimes, and foundational tools

# === LangChain Plugins / Integrations ===
langchain-openai       # Official integration for OpenAI models (ChatGPT, GPT-4, embeddings, etc.)
langchain-community    # Community-contributed integrations: tools, vector stores, retrievers, and more

# === Dev Tools & Deployment ===
langsmith              # Observability platform: log, trace, and debug LLM apps with full visibility
fastapi                # FastAPI framework for building APIs
python-dotenv          # Load environment variables from .env file

# Agent frameworks
langgraph
typing
langchain-tavily

# GenAI integration
langchain-google-genai
langchain_community
google-generativeai

# Vector DBs
langchain_chroma

# LangChain core
langchain-text-splitters

# Utilities
tqdm           # Hiển thị thanh progress khi indexing

# Hugging Face Embeddings
sentence-transformers
langchain-huggingface

# Dependencies cho custom embedding
transformers
torch

pydantic

# ==================== CHATBOT_THREAD1 ========================

#langchain==0.3.27
#langchain-core==0.3.79
#langchain-text-splitters==0.3.11

#langchain-google-genai==2.0.10
#google-generativeai==0.8.5

chromadb==1.3.4

tavily-python==0.7.13

#pydantic==2.12.4
pydantic-core==2.41.5

# =============================================================